Change Message#getMessageId() to a long instead of an int to futureproof for chats which have a lot of messages

+++ src/main/java/pro/zackpollard/telegrambot/api/chat/message/Message.java
@@ -17,7 +17,7 @@ public interface Message {
      *
      * @return The ID of the message
      */
-    int getMessageId();
+    long getMessageId();
 
     /**
      * Get the UNIX timestamp of when this message was sent/received at
